Move selector unification code to separate file
