+++ Makefile.conf
@@ -14,6 +14,7 @@ SOURCES = \
 	color_maps.cpp \
 	environment.cpp \
 	ast_sel_cmp.cpp \
+	ast_sel_unify.cpp \
 	ast_fwd_decl.cpp \
 	bind.cpp \
 	file.cpp \
