+++ win/libsass.vcxproj.filters
@@ -221,6 +221,9 @@
     <ClCompile Include="$(LIBSASS_SRC_DIR)\ast_sel_cmp.cpp">
       <Filter>Sources</Filter>
     </ClCompile>
+    <ClCompile Include="$(LIBSASS_SRC_DIR)\ast_sel_unify.cpp">
+      <Filter>Sources</Filter>
+    </ClCompile>
     <ClCompile Include="$(LIBSASS_SRC_DIR)\ast_fwd_decl.cpp">
       <Filter>Sources</Filter>
     </ClCompile>
